PR-B-7: extract StatementBinder (per-statement-kind binding)
Extract all 'Bind*Statement' methods from Binder.cs into a new
internal sealed class StatementBinder. This is the seventh of nine
extractions in the Phase-1 Binder decomposition.

- New: src/Core/CodeAnalysis/Binding/StatementBinder.cs (~2,275 LoC)
- Binder.cs shrinks from 14,243 to 12,123 LoC (-2,120 lines net)
- No public/internal API renames; all 92 RefactoringBaseline samples
  emit byte-identical IL (zero hash diffs).
- Constructor-injected dependencies: BinderContext, ConversionClassifier,
  PatternBinder, plus Func/delegate callbacks for the expression-binding,
  type-clause, attribute-binding, and symbol-construction helpers that
  remain on Binder.

Methods moved (54 in total):

Control-flow basics:
  BindStatement (dispatch), BindBlockStatement, BindBlockStatements,
  BindErrorStatement, BindVariableDeclaration (statement form),
  BindRefAliasLocalDeclaration, BindTupleDeconstructionStatement,
  BindNamedDeconstructionStatement, BindMultiAssignmentStatement

If/narrowing/switch:
  BindIfStatement, BindStatementWithNarrowing, BindSwitchStatement,
  ApplyMemberNotNullNarrowings, NarrowFieldIfNullable,
  InvalidateNarrowingsForAssignedVariables, CollectAssignedNames,
  TryClassifyNilGuard, TryClassifyBoolCallNarrowing,
  TryClassifyPatternNarrowing, IsNilLiteral, ClassifyImportedBoolCallNarrowing,
  ClassifyImportedMethodBoolCallNarrowing, ClassifyUserBoolCallNarrowing,
  MergeClrMemberNotNullWhenNarrowings, MergeUserMemberNotNullWhenNarrowings

Loops:
  BindForInfiniteStatement, BindForEllipsisStatement, BindForRangeStatement,
  BindForConditionStatement, BindForClauseStatement, BindLoopBody,
  BindAwaitForRangeStatement, GetIteratorElementType

Effects:
  BindTryStatement, BindThrowStatement, BindYieldStatement,
  BindUsingStatement, BindUsingStatementInBlock,
  BindDeferStatement, BindDeferStatementInBlock, IsDeferableCall,
  CaptureDeferArguments, CaptureArguments,
  BindGoStatement, BindChannelSendStatement, BindSelectStatement,
  BindScopeStatement, BuildCleanupTryStatement, ResolveExceptionType

Break/Continue/Return:
  BindBreakStatement, BindContinueStatement, BindReturnStatement,
  IsLvalueForRefReturn, HasFunctionLocalRefScope, HasFunctionLocalEscapeScope

Expression statement: BindExpressionStatement

Methods deliberately left on Binder:
  - 'BindVariableDeclaration(SyntaxToken, bool, TypeSymbol)' and the
    4-arg overload: symbol-construction helpers used in non-statement
    contexts (parameter binding, declaration binding); they will likely
    move to DeclarationBinder in PR-B-8.
  - 'BindSwitchExpression' and 'BindExpressionWithNarrowing': these are
    expressions, not statements; they belong in B-9 ExpressionBinder.
  - 'BindExpression' / 'BindTypeClause' / 'BindAttributes' /
    'IsLvalue' / 'BindVariableReference': injected as callbacks; remain
    the root Binder's responsibility until B-9.
  - 'IsIteratorReturnType' / 'IsAsyncIteratorReturnType' /
    'IsAsyncSequenceReturnType' / 'IsTaskShapedReturn': used by both
    statement-binding and declaration/method-signature contexts; remain
    utilities on Binder, with 'isIteratorReturnType' injected as a
    callback.

Note: open bug #568 ('using let x = expr' rejecting G# classes that
implement IDisposable) is intentionally NOT fixed in this PR. This PR
only sets up the structural home; the fix lands in StatementBinder
after the decomposition completes.

Next PR in the sequence: B-8 DeclarationBinder.
